backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.6",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
]
